[FIX][ADD] setting auto_set_title
bash source the file /etc/bash.bashrc before reading any other configuration
in this file PROMPT_COMMAND is set to update the window title to: USER@HOST: PWD
which will override i3terms custom window title (path to the terminal /dev/pts/X)

which in turn made --palette-menu option not work, since it needs the path in the title

this commit adds the configuration setting "auto_set_title" if it is set to "true"
and the shell is bash, and the command to execute is bash (i.e. no command)
we will do this: bash --rcfile <(echo unset PROMPT_COMMAND; cat ~/.bashrc)
so that the window title does not update. It is unlikely that the PROMPT_COMMAND
is set to anything other than update the title at this point, and the user
can still customize the PROMPT_COMMAND in their .bashrc. One side effect
is that the path to the rcfile is not "~/.bashrc" which could also possible
break some obscure hacks where that is relevant. so if this fix casue any
issues, it can simply be disabled from the config.